Fold ol.CoordinateFormat into ol.Coordinate
